validations: Fix CPF validation logic
Previous logic didn't account for the 1st check digit result being
exactly `10`.

According to spec, the result of the `modulo(11)` operation should be
set to `0` if it results in a `10`, before adding re-applying the mask
multipliers for the second pass (including with the new `0` at the end
of the numbers list).

Fixes #194